+++ Semana 1/react/src/soma.js
@@ -1,8 +1,0 @@
-const soma = (a, b) => {
-    return a + b;
-  };
-
-
-module.exports ={
-    soma
-}